+++ src/lib/pages/validator-details/components/validator-overview/index.tsx
@@ -70,12 +70,13 @@ export const ValidatorOverview = ({
       ) : (
         <>
           <Flex backgroundColor="gray.900" p={6} rounded={8} w="100%">
-            <RecentBlocksSection hasTitle />
+            <RecentBlocksSection validatorAddress={validatorAddress} />
           </Flex>
           <VotingPowerChart
             validatorAddress={validatorAddress}
             singleStakingDenom={singleStakingDenom}
             assetInfos={assetInfos}
+            isOverview
           />
         </>
       )}